@@ -18,19 +18,16 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  simpleprogressdialog: ^1.0.8
   image_pixels: ^2.0.0
   dio: ^4.0.0
   image: ^3.0.2
   image_picker: ^0.7.3
   image_cropper: ^1.4.0
-  image_gallery_saver: '^1.6.8'
   path_provider: ^2.0.1
   permission_handler: ^6.1.1
-  flushbar: ^1.10.4
   flutter:
     sdk: flutter
 
@@ -49,11 +46,11 @@ dev_dependencies:
 
 # The following section is specific to Flutter.
 flutter:
+  uses-material-design: true
 
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
-  uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
   # assets:
@@ -89,6 +86,10 @@ flutter:
     - family: Righteous
       fonts:
         - asset: fonts/Righteous-Regular.ttf
+    - family: CustomIcon
+      fonts:
+        - asset: fonts/CustomIcon.ttf
+
 
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
